[#5] Using simple math properties.
This does not teach you anything about scala but should probably be really fast.

Some explanation: let's decompose the sum of the suite of multiples of 3 below 1000:
```3 + 6 + 9 + 12 + ... + 999```
If we divide everything by 3, this can be rewritten as:
```3 * (1 + 2 + 3 + 4 + ... + 333)```
The sum of such an integer suite is simple to compute:
```sum(1..n) = n * n+1 / 2```
Now we can rewrite the suite of multiple of 3 below 1000 as:
```333 * (333 + 1) /2```
Simple no ? Let's call this value 'sum3'

Now, we can do the same with multiples of 5. Let's call that result 'sum5'.

If you add sum3 and sum5, you will not find the expected value because multiples of 3 and 5 (ie multiples of 15) are taken two times into account.

So, let's compute the sum of multiples of 15 below 1000 and call that variable sum15.

Our final result will be: sum3 + sum5 - sum15.